+++ src/main.jsx
@@ -1,10 +1,13 @@
 import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
+import ReactDOM from "react-dom";
+import App from "./App";
 import "./index.css";
-
-ReactDOM.createRoot(document.getElementById("root")).render(
+import { BrowserRouter} from "react-router-dom";
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+    <BrowserRouter>
+    <App/>
+    </BrowserRouter>
+  </React.StrictMode>
 );
